major update, removed old code which was not used anymore, dictionaries, spreadsheet, and tokenizer are now gone
