+++ lingpy/algorithm/_tree.py
@@ -1,12 +1,3 @@
-'''
-Created on Feb 28, 2012
-
-@author: PhyloStar
-Computes the Generalized RF distance, RF-distance and Normalized RF-distance
-
-Modified and adapted to LingPy style on 2014-08-15 by Johann-Mattis List
-'''
-
 import re
 from ..thirdparty import LoadTree as Tree
