--- lingpy/tests/meaning/__init__.py
@@ -1,7 +1,0 @@
-# author   : Peter Bouda
-# email    : pbouda@cidles.eu
-# created  : 2013-08-26 10:00
-
-
-__author__="Peter Bouda"
-__date__="2013-08-26"
